fix 64-bit varint truncation in decode_fast

+++ upb/wire/decode_fast/cardinality.h
@@ -440,9 +440,9 @@ bool upb_DecodeFast_Unpacked(upb_Decoder* d, const char** ptr, upb_Message* msg,
 UPB_FORCEINLINE bool _upb_DecodeFast_DecodeSizeSlow(const char** pp,
                                                     int* size) {
   const char* ptr = *pp;
-  uint32_t val = (ptr[0] & 0x7f) | ((ptr[1] & 0x7f) << 7);
+  uint64_t val = (ptr[0] & 0x7f) | ((ptr[1] & 0x7f) << 7);
   for (int i = 2; i < 5; i++) {
-    uint32_t byte = (uint8_t)ptr[i];
+    uint64_t byte = (uint8_t)ptr[i];
     val |= (byte & 0x7f) << (i * 7);
     if (!(byte & 0x80)) {
       if (UPB_UNLIKELY(val > INT32_MAX)) return false;
